Bump the default search depth of git.Tag.latest()
0.6.3

Even on a huge repo like v8, a --max-count=100 still runs quite
fast on my machine, and if you have a lot of commits since your
last release tag, 10 might not be enough to find it. We may
also consider falling back to a single `git describe --exact`
call in case our searchDepth couldn't find it.